requirements.txt
aniso8601==8.0.0
apispec==2.0.2
appdirs==1.4.4
attrs==20.2.0
black==18.9b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssmin==0.2.0
distlib==0.3.1
Faker==4.1.2
filelock==3.0.12
Flask==1.1.2
flask-apispec==0.8.2
Flask-Assets==2.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lesscpy==0.14.0
MarkupSafe==1.1.1
marshmallow==2.21.0
pipenv==2020.8.13
ply==3.11
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
webargs==6.1.1
webassets==2.0
Werkzeug==1.0.1
zipp==3.1.0